migrated Firebase deployment, Pinecone, and Hugging Face to ktp lambda tech dev email for chatbot

+++ backend/assistant/update_document_store.py
@@ -1,9 +1,11 @@
+import argparse
 from dotenv import load_dotenv
 from langchain_huggingface import HuggingFaceEmbeddings
 from langchain.text_splitter import RecursiveCharacterTextSplitter
 import os
 import pandas as pd
 from pinecone import Pinecone
+from typing import Sequence
 
 load_dotenv()
 
@@ -20,14 +22,14 @@ def get_info() -> pd.DataFrame:
     return info
 
 
-def create_document_embeddings(info: pd.DataFrame) -> list[dict]:
+def create_document_embeddings(info: pd.DataFrame) -> Sequence[dict]:
     """Converts the source data into vector embeddings for Pinecone.
 
     Args:
         info (pd.DataFrame): the source data, formatted as a dataframe with columns "id", "sources", and "metadata".
 
     Returns:
-        records (list[dict]): the vector embeddings of the source data.
+        records (Sequence[dict]): the vector embeddings of the source data.
     """
 
     # Chunks the source data
@@ -73,36 +75,54 @@ def create_document_embeddings(info: pd.DataFrame) -> list[dict]:
     return records
 
 
-def update_pinecone(records: list, pc: Pinecone):
+def update_pinecone(records: Sequence, pc: Pinecone, insert_only: bool = False):
     """Updates the vector embeddings in Pinecone.
 
     Args:
-        records (list): the vector embeddings of the source data.
+        records (Sequence): the vector embeddings of the source data.
         pc: (Pinecone): an instance of the Pinecone class.
+        insert_only (bool): specifies whether existing vectors should be deleted.
 
     Returns:
         None: no return value.
     """
 
+    # Connects to Pinecone
     try:
-        # Connects to Pinecone
         index = pc.Index(host=os.getenv("PINECONE_HOST"))
+    except Exception as e:
+        print("Failed to connect to Pinecone")
+        raise e
 
-        # Deletes the old vector embeddings in Pinecone
-        index.delete(delete_all=True)
-        print("Successfully deleted old vector embeddings in Pinecone")
+    # Deletes the old vector embeddings in Pinecone
+    try:
+        if not insert_only:
+            index.delete(delete_all=True)
+            print("Successfully deleted old vector embeddings in Pinecone")
+    except Exception as e:
+        print(
+            "Failed to delete old vector embeddings in Pinecone (if old embeddings do not exist, run the script with the -i flag)"
+        )
+        raise e
 
-        # Upserts the new vector embeddings in Pinecone
+    # Upserts the new vector embeddings in Pinecone
+    try:
         upsert_response = index.upsert(vectors=records)
         print(
             f'Successfully upserted {upsert_response["upserted_count"]} vector embeddings in Pinecone'
         )
     except Exception as e:
-        print(f"Failed to update vector embeddings in Pinecone: {e}")
+        print(f"Failed to upsert vector embeddings in Pinecone: {e}")
         raise e
 
 
 if __name__ == "__main__":
+
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "-i", "--insert-only", help="Only insert new vectors", action="store_true"
+    )
+    args = parser.parse_args()
 
     # Reads the document source data
     info = get_info()
@@ -112,4 +132,4 @@ if __name__ == "__main__":
     records = create_document_embeddings(info=info)
 
     # Updates the vector embeddings in Pinecone
-    update_pinecone(records=records, pc=pc)
+    update_pinecone(records=records, pc=pc, insert_only=args.insert_only)
